Revert "Merge master into teravuswork", it should have been avination, not master.
This reverts commit dfac269032300872c4d0dc507f4f9062d102b0f4, reversing
changes made to 619c39e5144f15aca129d6d999bcc5c34133ee64.

+++ OpenSim/Services/AssetService/XAssetService.cs
@@ -194,7 +194,21 @@ namespace OpenSim.Services.AssetService
             if (!UUID.TryParse(id, out assetID))
                 return false;
 
-            return m_Database.Delete(id);
+            AssetBase asset = m_Database.GetAsset(assetID);
+            if (asset == null)
+                return false;
+
+            if ((int)(asset.Flags & AssetFlags.Maptile) != 0)
+            {
+                return m_Database.Delete(id);
+            }
+            else
+            {
+                m_log.DebugFormat("[XASSET SERVICE]: Request to delete asset {0}, but flags are not Maptile", id);
+            }
+
+            return false;
         }
     }
 }
+
